requirements.txt
google-auth
google-auth-oauthlib
packaging
google-api-python-client
progressbar2